feat: enhance seat icon functionality with viewBox normalization and rendering improvements
